Move tests to separate file as requested
